GSA_track1_cat1/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.4.0